improve grammar live collection
